Version 1
This is the first version of the code. There is a basic outline of how the project is going to look. In this state, it is very messy and unoptimised. There are currently only 2 functions in this code that take information from the database. It offers insight into who the top 15 artists were on Spotify back in August 2023.