Version 2
This is the second version of the project. There is now 4 functions that users can use to receive the data. I imported os and time. The data is now displayed on a much nicer format and the code has been optimized to reduce clunkiness. I also added a new column "Gender" to my database. There were a lot of issues when formatting my data output because of the length of the names and countries, though I was able to fix this using len().