Version 3
This is the third version of this project. The main change of this version is that each function has an ascending or descending option. This is convenient for the user as they can reverse the order of a filter with no trouble. I also added .strip() and .lower() on the necessary input codes to make it easier for the user. Lastly, I changed the wording for some things as it was not so clear what they meant.