Version 4
This is the final version of my project. In this version, I have added 2 more functions. The first allows the user to get the average of some of the statistics such as monthly listeners and albums released. It adds up the value from all 15 artists and divides it by 15, then giving a statement to the user what it is. The second function is  an advanced delete function. It allows the user to delete an artist from the database so the data table will not display them anymore. This can create lots of custom tables, and also changes the average. Another improvement of this version is the added comments in my code. This makes it easier to understand what parts of my code do what.